fix: move aws config to lib
