+++ src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './aws'
 export * from './capitalize'
 export * from './createMockRouter'
 export * from './getColorsForExport'
